requirements.txt
python-dotenv
requests
psycopg2-binary
numpy
matplotlib
scikit-learn
pandas
pymongo